Created new macros and correspondence files
Hi Rob, I tried to create the new Macros, shapefiles, and correspondence files, but I think I might've done something wrong. When I ran Block00_npa19 in the prog folder, all the NPA values in the results viewer were PRE. Is this because I haven't batch submitted yet? Also, there were instructions to save some of the sas stat transfer files to folders I don't have on my C drive (ie the maps subfolder in the general folder) not sure if that has anything to do with it.